Use 'Instances' for sending NOTIFYs on publication
